scheme: base classes are now in (nomad platform api)
By using (nomad platform api) we ca have it so only generic code is used in
nomad in general. This allows us to keep modules organized in a way that does
not cause cyclical cycles.

The platform classes are no longer prefixed with platform. We can now use nice
naming conventions for end users an regular nomad extension.