guix: nomad-git declaration now uses 'g-golf-git as it's input
This fixes an issue where nomad-git and g-golf version could be out of sync.

guix environment -L ./guix nomad-git now uses the g-golf-git
input. This ensures we can the pin the "g-golf" input to a certain g-golf
commit. Useful if we are using bug fixes or features that are not quite in
Guix' g-golf package.