views: root-view now describes interactive commands.
This adds a table field with the description of the interactive command bound
to a key. This dynamically produced from the interactive command docstring.

This also ensures all %web-mode-map commands have doc strings and improves on
existing ones.

The interactive command 'home now will open URI nomad: if %default-home-page is
\#f. This ensure that home will also do something useful.