doc: new procedure doc->shtml. Converts doc-string texinfo to shtml,
When using HTML views, doc->shtml is used to convert doc-strings with texinfo
markup to shtml expressions. This ensure the texinfo is stripped but still
presentable to the user in HTML.

Also now root-view sorts web-mode-map key map by it's keys. This makes the
output more predictable and logically ordered.

Views defined with 'define-view now have there links open in a new
buffer.