menu: URI minibuffer completing read now users #:initial-input
This ensure that edit-menu-uri works much the same as #:edit-uri.  Where the
current URI is populated into the editing minibuffer.

The GtkEntry now uses the #:overwrite-mode property. This mode uses a square
cursor which is more emacs like. And on par with the default
minibuffer. Since we don't actually use key press events at the GtkEntry level
the overwrite part of the mode is irrelevant.

The 'focus-in-event signal now calls 'edit-menu-uri though by emulating C-c
e. This is a short term hack until we can resolve some emacsy co routine
issues.

Clean up of the minibuffer prompts and comments now occurs on buffer exit.