menu: edit-menu-uri now also handles query.
When editing the URI buffer if the input is not a valid URT it is treated as a query.

(nomad uri) now has a new procedure 'valid-uri? which validates a URI.

The interactive command 'query now has an optional argument buffer, when buffer is used a query is
performed with that buffer. Alternatively a new buffer is created.