web-mode: C-o 'load-uri is now C-c l 'load-uri
This fixes an issue when using emacsy's default C-o 'other-buffer which is a
emacsy glabal-map. Since windows are only partially implemented we don't
actually support 'other-buffer.  And using C-o in %web-mode-map would shadow
this and fail silently in non web buffers. Potentially causing GTK widget
issues.

C-o is now redefined to use 'undefined-command and 'undefined-command has been
extended to take a command string so users know which interactive command is
undefined. Later we should add a unimplemented-command to differentiate
between a command that is not implemented yet and a command that might not be
defined.

Since emacs C-o is 'other-line then we indicate this is currently
undefined. Since 'other-line does not exist in emacsy at all.