web-mode: remove unused 'load-clipboard-uri command.
'load-clipboard-uri is not used and implies a URI is being loaded, but in
reality it was creating a buffer.

Also despite adding 'load-uri to %gi-method-short-names-skip. g-golf still
tries to use the short method which conflicts with the 'load-uri command. So
'load-clipboard-uri is broken. And g-golf has a potential bug here.

This removes the 'load-clipboard-uri to avoid the g-golf bug and any confusion
with 'make-buffer-clipboard which should be used instead.