ibuffer: now uses line index instead of a slot state index.
Previously we did not have enough emacsy primitives to handle line index so we
had to maintain an <ibuffer> !index slot for state.  This meant we had to
synchronize lines and states which complicated the code and created bugs.

Now we use line index and reference buffers in the !buffers slot. This ensure
state is more accurately maintained.

We also now reverse the 'buffer-list this mean buffers are listed in the order
they are created, this also helps with line indexing.  This may not exactly
conform to emacs default ibuffer behavior. But it simplifies the code.

3 new private module bindings have been added. 'header-lines, 'hidden-buffer,
and 'line-offset. These just make sure the line-offset is consistent and not
hard coded.

ibuffer now also has interactive tests. This should ensure that we have
expected behavior and avoid any regressions in the future.