web-mode: <web-buffer> now uses buffer-modes instead of local-keymap
This fixes an issue where the keymap for web buffers were not actually using a
<mode> and only using a local-keymap. Making it hard to introspect the
<web-buffer> buffer modes.

%firefox-webview-map is now renamed to %firefox-mode-map.

%firefox-mode-map and %web-mode-map are now constructed at the binding
definition. And instead of using list->keymap which has been removed. The new
define-each-key is used to add key/command pairs to the appropriate keymap.

<web-buffer> now constructs the buffer-modes using #:init-form this ensures
that the #:mode-map for the 'web-mode is evaluated when the buffer is
constructed.

The web-mode binding is now a procedure this ensures that %web-mode-map is
evaluated whenever a web-mode is constructed.