doc: new class <doc> and new procedure find-doc
The new procedure 'find-doc is used to find documentation for a symbol. It can
returns a list of <doc> or a <doc>.

This is a preliminary commit. Though describe-function now outputs more
details about a procedure or function. It does not use 'find-doc which is still
work in process.

The new procedure 'doc-loc is more work in process and will eventually find
the document location for a symbol.