uri: '<uri-handler> now handles documentation links.
This provided support for loading links with custom URI scheme. Previously
there was not enough context to load the a URI without first makeing a new
'<help-buffer> with an interactive command.

This also allows custom URI scheme to be loaded directly using a URI. For
example use 'load-uri we can now describe an object with M-x: load-uri <RET>
nomad:/describe/object/%load-path.

'<nomad-uri-handler> now no longer uses a uri slot. The URI is passed as a
arguement to the handler methods instead.

The new '%request-uri parameter is now used to provided context for view
paths. Since views are rendered before a web view is created there is no way
to get the 'current-url, So we must handle this contextually. This is an
interim solution until we can pass URI's as a procedure argument. Which means
reworking 'define-view to handle that.

The interactive command 'rename-buffer now takes the optional argument buffer
last. This makes the procedure more useful when being called none
interactively.

The interactive command 'ensure-uri now ignores the nomad custom URI. Since it
 does not have a URI protocol there is no point prefixing it with https making
 easier now to load nomad URI's using 'load-uri