uri: '<nomad-handle-uri>'s handle-uri is now defined in help-mode
This ensures we don't have cyclical module usage when checking for the current
buffer's class.

When handling nomad: URI if the current buffer we are loading the URI with, is
not a <help-buffer> we now change it to <help-buffer>. This ensures web
buffers that are not instantiated with make but load a nomad: URI directly
gain help mode. And 'describe-mode will now accurately display buffer's
intended mode.